% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equality.R
\name{are_equal_f}
\alias{are_equal_f}
\title{equals}
\usage{
are_equal_f(x, y, eps = 1)
}
\arguments{
\item{x}{LHS}

\item{y}{RHS}

\item{eps}{Precision of floating-point comparison}
}
\value{
Logical vector
}
\description{
equals
}
