% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floats.R
\name{are_same_all}
\alias{are_same_all}
\title{are_same_all}
\usage{
are_same_all(x, eps = 1)
}
\arguments{
\item{x}{Vector to test}

\item{eps}{Floating-point comparison tolerance}
}
\value{
\code{TRUE} if all elements are equal
}
\description{
are_same_all
}
