% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asserts.R
\name{assert}
\alias{assert}
\title{Assert}
\usage{
assert(expr, msg = NULL, subclass = NULL)
}
\arguments{
\item{expr}{Expression to test. Should evaluate to logical.}

\item{msg}{Optional message to include in the error.}

\item{subclass}{Additional error subclasses.}
}
\value{
Invisibly \code{TRUE} if successful.
}
\description{
Assert
}
