% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{cc}
\alias{cc}
\title{Vec_c}
\usage{
cc(
  ...,
  .ptype = NULL,
  .name_spec = NULL,
  .name_repair = c("minimal", "unique", "check_unique", "universal")
)
}
\arguments{
\item{...}{Items to concatenate.}

\item{.ptype, .name_spec, .name_repair}{Passed to \code{vctrs::vec_c}.}
}
\description{
A shortcut for \code{vctrs::vec_c}.
}
