% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconstruction.R
\name{\%->\%}
\alias{\%->\%}
\alias{\%<-\%}
\title{Deconstruction operators}
\usage{
lhs \%->\% rhs

lhs \%<-\% rhs
}
\arguments{
\item{lhs, rhs}{Left- and right-hand side of the operator}
}
\value{
Data (invisibly)
}
\description{
Mimicks \code{zeallot}'s beahviour
}
