% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floats.R
\name{distinct_which_f}
\alias{distinct_which_f}
\title{distinct_which_f}
\usage{
distinct_which_f(x, y, eps = 1L)
}
\arguments{
\item{x}{First vector.}

\item{y}{Second vector.}

\item{eps}{Floating-point comparison tolerance}
}
\value{
List with two items, \code{x} and \code{y}, which contain indices of values that are not present in another collection
}
\description{
distinct_which_f
}
