% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equality.R
\name{\%==\%}
\alias{\%==\%}
\alias{\%==\%.double}
\alias{\%==\%.name}
\alias{\%==\%.default}
\alias{\%==\%.default.default}
\alias{\%==\%.default.double}
\alias{\%==\%.default.name}
\alias{\%!=\%}
\alias{\%===\%}
\alias{\%!==\%}
\title{Equality operators}
\usage{
x \%==\% y

\method{\%==\%}{double}(x, y)

\method{\%==\%}{name}(x, y)

\method{\%==\%}{default}(x, y)

\method{\%==\%.default}{default}(x, y)

\method{\%==\%.default}{double}(x, y)

\method{\%==\%.default}{name}(x, y)

x \%!=\% y

x \%===\% y

x \%!==\% y
}
\arguments{
\item{x}{Lhs.}

\item{y}{Rhs.}
}
\value{
Either a logical vector of the same as input size, or a single logical value.
}
\description{
Equality operators
}
