% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_fmt.R
\name{glue_fmt}
\alias{glue_fmt}
\alias{glue_fmt_chr}
\title{String interpolation}
\usage{
glue_fmt(..., .envir = parent.frame())

glue_fmt_chr(..., .envir = parent.frame())
}
\arguments{
\item{...}{Parameters passed to \code{glue::glue}.}

\item{.envir}{Evaluation environment.}
}
\value{
Interpolated \code{glue} strings.
}
\description{
String interpolation
}
