% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%&\%}
\alias{\%&\%}
\title{Concat/add infix operator.}
\usage{
x \%&\% y
}
\arguments{
\item{x}{Left summand.}

\item{y}{Right summand.}
}
\value{
Result of the aapropriate summation/concatenation.
}
\description{
Performs (possibly) a vectorized summation operation,
 which depends on the class of operators.
 Following methods are implemented:
 \code{character} + \code{character},
       1-to-1 vectorized, concatenation of strings.
Does the same as `%+%`.
}
