% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{len}
\alias{len}
\alias{len.default}
\alias{len.unit}
\alias{len.quosures}
\title{len}
\usage{
len(x)

\method{len}{default}(x)

\method{len}{unit}(x)

\method{len}{quosures}(x)
}
\arguments{
\item{x}{Object to measure.}
}
\value{
Length of the object.
}
\description{
len
}
