% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{lin}
\alias{lin}
\title{lin}
\usage{
lin(x, x0, y0)
}
\arguments{
\item{x}{Where to interpolate.}

\item{x0}{Arguments (size 2).}

\item{y0}{Values (size 2).}
}
\value{
Interpolated value between two provided.
}
\description{
lin
}
