% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floats.R
\name{unique_which_f}
\alias{unique_which_f}
\title{unique_which_f}
\usage{
unique_which_f(x, eps = 1L)
}
\arguments{
\item{x}{Vector to test.}

\item{eps}{Floating-point comparison tolerance}
}
\value{
Indices of items that are unique within the source vector.
}
\description{
unique_which_f
}
