% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{vec_names}
\alias{vec_names}
\alias{vec_names.data.frame}
\alias{vec_names.default}
\title{Vector names}
\usage{
vec_names(x, ...)

\method{vec_names}{data.frame}(x, ...)

\method{vec_names}{default}(x, ...)
}
\arguments{
\item{x}{Vector to get names from.}

\item{...}{Placeholder parameters.}
}
\value{
Names of the vector.
}
\description{
Vector names
}
