% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{vec_rips}
\alias{vec_rips}
\alias{vec_rip}
\title{Vector accessor}
\usage{
vec_rips(x, i)

vec_rip(x, i)
}
\arguments{
\item{x}{Vector to slice.}

\item{i}{Indexes.}
}
\value{
\code{list_of<item_ptype>}.
}
\description{
Temporary solution.
}
