% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%win\%}
\alias{\%win\%}
\alias{\%wini\%}
\alias{\%!win\%}
\alias{\%!wini\%}
\title{Within}
\usage{
lhs \%win\% rhs

lhs \%wini\% rhs

lhs \%!win\% rhs

lhs \%!wini\% rhs
}
\arguments{
\item{lhs}{Numeric vector to test}

\item{rhs}{A pair of boundaries}
}
\value{
Logical vector, indicating the position of items within the range
}
\description{
Within
}
