% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{as_list_of.default}
\alias{as_list_of.default}
\alias{as_list_of.data.frame}
\alias{as_vec}
\title{As converters}
\usage{
\method{as_list_of}{default}(x, ..., .ptype = NULL)

\method{as_list_of}{data.frame}(x, ..., .ptype = NULL)

as_vec(x, ...)
}
\arguments{
\item{x}{Collection to convert.}

\item{...}{Unused args, for compatibility.}

\item{.ptype}{Fallback type.}
}
\value{
\code{list_of}/\code{vector} as a result of conversion.

\code{list_of}/\code{vector} as a result of conversion.
}
\description{
As converters

As converters
}
