% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equality.R
\name{\%==\%}
\alias{\%==\%}
\alias{\%==\%.double}
\alias{\%==\%.name}
\alias{\%==\%.default}
\alias{\%==\%.default.default}
\alias{\%==\%.default.double}
\alias{\%==\%.default.name}
\alias{\%!=\%}
\alias{\%===\%}
\alias{\%!==\%}
\title{Equality operators}
\usage{
`x \%==\% y`

`(0.1 + 0.2) \%==\% (0.3)`

`name_1 \%==\% name_2`

`1.0 \%==\% y`

`1 \%==\% "1"`

`"1" \%==\% 1.0`

`"x" \%==\% x`

x \%!=\% y

x \%===\% y

x \%!==\% y
}
\arguments{
\item{x}{Lhs.}

\item{y}{Rhs.}
}
\value{
Either a logical vector of the same as input size, or a single logical value.
}
\description{
Equality operators
}
