% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{vmap}
\alias{vmap}
\alias{vmap_pt}
\alias{vmap2}
\alias{vmap2_pt}
\alias{vmap_if}
\alias{vmap_at}
\title{Mappers}
\usage{
vmap(.x, .f, ..., .fb_ptype = NULL)

vmap_pt(.x, .f, ..., .fb_ptype = NULL)

vmap2(.x, .y, .f, ..., .fb_ptype = NULL)

vmap2_pt(.x, .y, .f, ..., .fb_ptype = NULL)

vmap_if(.x, .p, .f, ..., .else = NULL, .fb_ptype = NULL)

vmap_at(.x, .at, .f, ..., .fb_ptype = NULL)
}
\arguments{
\item{.x, .y}{Vectors to iterate over.}

\item{.f}{Mapper function.}

\item{...}{Additional paramters passed to mappers}

\item{.fb_ptype}{Fallback \code{ptype} for type-stability of empty collections.}

\item{.p}{Predicate function.}

\item{.else}{Alternative mapper for \code{vmap_if}.}

\item{.at}{Location to map at.}
}
\value{
Either \code{list_of} or vector of common \code{ptype}.
}
\description{
Performs \code{vctrs} - compatibel mapping.
Requries strongly typed input and output. By default returns \code{vctrs::list_of}.
\code{*_pt} version performs conversion to the common \code{ptype}.
}
